move counts to separate worker
