+++ manifest.json
@@ -199,6 +199,14 @@
       "group": "backend",
       "exec_mode": "cluster",
       "instances": 1
+    },
+    {
+      "script": "svc/counts.js",
+      "watch": true,
+      "ignore_watch": [".git", "node_modules"],
+      "group": "backend",
+      "exec_mode": "cluster",
+      "instances": 1
     }
   ]
 }
